+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Movim Changelog
 v0.17.1 (trunk)
 ---------------------------
  * Improve Upload widget, add drag & drop feature
+ * Remove the custom CSS blog feature
 
 v0.17
 ---------------------------